divide the contract into different files
